Server is now communicating with JSON-s instead of raw Java Objects | ArrayList -> List | InterstellarWar game objects type refactor | junit test improves (now they should work)

+++ src/net/neczpal/interstellarwar/common/game/Planet.java
@@ -4,52 +4,52 @@ import java.io.Serializable;
 import java.util.ArrayList;
 
 public class Planet implements Serializable {
-	public static final int PLANET_TYPES = 18;
-	private static final long serialVersionUID = 2683452581122892189L;
+    public static final int PLANET_TYPES = 18;
+    private static final long serialVersionUID = 2683452581122892189L;
 
-	private float mX;
-	private float mY;
-	private float mRadius;
+    private int mX;
+    private int mY;
+    private int mRadius;
 
-	private int mTextureIndex;
+    private int mTextureIndex;
 
-	private int mOwnedBy;
-	private int mUnitsNumber;
-	private ArrayList <Planet> mNeighbors;
+    private int mOwnedBy;
+    private int mUnitsNumber;
+    private ArrayList<Planet> mNeighbors;
 
 
-	/**
-	 * Erstellt einen Planet
-	 *
-	 * @param x           Position auf dem X-Achse
-	 * @param y           Position auf dem Y-Achse
-	 * @param radius      Radius von dem Planet
-	 * @param ownedBy     Die Nummer von Spieler, die diesem Planet dominiert
-	 * @param unitsNumber Die Anzahl der Einheit auf dem Planet
-	 */
-	Planet (float x, float y, float radius, int ownedBy, int unitsNumber) {
-		this (x, y, radius, ownedBy, unitsNumber, (int) (Math.random () * PLANET_TYPES));
-	}
+    /**
+     * Erstellt einen Planet
+     *
+     * @param x           Position auf dem X-Achse
+     * @param y           Position auf dem Y-Achse
+     * @param radius      Radius von dem Planet
+     * @param ownedBy     Die Nummer von Spieler, die diesem Planet dominiert
+     * @param unitsNumber Die Anzahl der Einheit auf dem Planet
+     */
+    Planet (int x, int y, int radius, int ownedBy, int unitsNumber) {
+        this (x, y, radius, ownedBy, unitsNumber, (int) (Math.random () * PLANET_TYPES));
+    }
 
-	/**
-	 * Erstellt einen Planet
-	 *
-	 * @param x           Position auf dem X-Achse
-	 * @param y           Position auf dem Y-Achse
-	 * @param radius      Radius von dem Planet
-	 * @param ownedBy     Die Nummer von Spieler, die diesem Planet dominiert
-	 * @param unitsNumber Die Anzahl der Einheit auf dem Planet
-	 * @param tex         Die Texture des Planets
-	 */
-	Planet (float x, float y, float radius, int ownedBy, int unitsNumber, int tex) {
-		mTextureIndex = tex;
+    /**
+     * Erstellt einen Planet
+     *
+     * @param x           Position auf dem X-Achse
+     * @param y           Position auf dem Y-Achse
+     * @param radius      Radius von dem Planet
+     * @param ownedBy     Die Nummer von Spieler, die diesem Planet dominiert
+     * @param unitsNumber Die Anzahl der Einheit auf dem Planet
+     * @param tex         Die Texture des Planets
+     */
+    Planet (int x, int y, int radius, int ownedBy, int unitsNumber, int tex) {
+        mTextureIndex = tex;
 
-		mX = x;
-		mY = y;
-		mRadius = radius;
-		mUnitsNumber = unitsNumber;
-		mOwnedBy = ownedBy;
-		mNeighbors = new ArrayList <> ();
+        mX = x;
+        mY = y;
+        mRadius = radius;
+        mUnitsNumber = unitsNumber;
+        mOwnedBy = ownedBy;
+        mNeighbors = new ArrayList <> ();
 	}
 
 	/**
@@ -99,42 +99,42 @@ public class Planet implements Serializable {
 	 * @param px Position auf dem X-Achse
 	 * @param py Position auf dem Y-Achse
 	 * @return Entscheidet ob ein Punkt (px, py) innerhalb ist
-	 */
-	public boolean isInside (float px, float py) {
-		return Math.pow (px - mX, 2) + Math.pow ((py - mY), 2) <= mRadius * mRadius;
-	}
+     */
+    public boolean isInside (float px, float py) {
+        return Math.pow (px - mX, 2) + Math.pow ((py - mY), 2) <= mRadius * mRadius;
+    }
 
-	/**
-	 * @param planet die andere Planet
-	 * @return Entscheidet ob die andere Planet Nachbar von diesem Planet ist
-	 */
-	public boolean isNeighbor (Planet planet) {
-		return mNeighbors.contains (planet);
-	}
+    /**
+     * @param planet die andere Planet
+     * @return Entscheidet ob die andere Planet Nachbar von diesem Planet ist
+     */
+    public boolean isNeighbor (Planet planet) {
+        return mNeighbors.contains (planet);
+    }
 
-	//SETTER, GETTERS
+    //SETTER, GETTERS
 
-	public float getX () {
-		return mX;
-	}
+    public int getX () {
+        return mX;
+    }
 
-	public float getY () {
-		return mY;
-	}
+    public int getY () {
+        return mY;
+    }
 
-	public float getRadius () {
-		return mRadius;
-	}
+    public int getRadius () {
+        return mRadius;
+    }
 
-	public int getOwnedBy () {
-		return mOwnedBy;
-	}
+    public int getOwnedBy () {
+        return mOwnedBy;
+    }
 
-	public int getUnitsNumber () {
-		return mUnitsNumber;
-	}
+    public int getUnitsNumber () {
+        return mUnitsNumber;
+    }
 
-	public void setUnitsNumber (int unitsNumber) {
+    public void setUnitsNumber (int unitsNumber) {
 		this.mUnitsNumber = unitsNumber;
 	}
